refactor(#83): move context menu to a file
