+++ src/app/index.ts
@@ -15,116 +15,7 @@ import { Background } from './Background';
 import { requestHandlers } from './Background/requestHandlers';
 
 import { sendConfigUpdateEvent } from './ContentScript';
-
-import browser from 'webextension-polyfill';
-import { getPageTranslateState } from '../requests/contentscript/pageTranslation/getPageTranslateState';
-import { createEvent, createStore } from 'effector';
-import { getTranslatorFeatures } from '../requests/backend/getTranslatorFeatures';
-import { getConfig } from '../requests/backend/getConfig';
-import { disableTranslatePage } from '../requests/contentscript/pageTranslation/disableTranslatePage';
-import { enableTranslatePage } from '../requests/contentscript/pageTranslation/enableTranslatePage';
-import { pageTranslatorStateUpdatedHandler } from './ContentScript/PageTranslator/requests/pageTranslatorStateUpdated';
-
-type PageTranslationState = {
-	tabId: null | number;
-	isTranslating: boolean;
-};
-
-// TODO: move to a file
-// TODO: add i18n texts
-// TODO: add method `stop`
-class TranslatePageContextMenu {
-	private menuItemId = 'translatePage';
-
-	private tabStateUpdated;
-	private $tabState;
-
-	constructor() {
-		this.tabStateUpdated = createEvent<PageTranslationState>();
-		this.$tabState = createStore<PageTranslationState>({
-			tabId: null,
-			isTranslating: false,
-		});
-
-		this.$tabState.on(this.tabStateUpdated, (_state, payload) => payload);
-	}
-
-	public start() {
-		browser.contextMenus.create({
-			id: this.menuItemId,
-			contexts: ['page'],
-			title: 'Translate page',
-			// type: 'checkbox',
-			// checked: false,
-		});
-
-		this.$tabState.watch((state) => {
-			browser.contextMenus.update(this.menuItemId, {
-				title: state.isTranslating ? 'Disable translation' : 'Translate page',
-			});
-		});
-
-		browser.contextMenus.onClicked.addListener(this.onClickMenu);
-
-		const updateMenuItem = async (tabId: number) => {
-			const translateState = await getPageTranslateState(tabId);
-			this.tabStateUpdated({
-				tabId,
-				isTranslating: translateState.isTranslated,
-			});
-		};
-
-		browser.tabs.onActivated.addListener((info) => {
-			updateMenuItem(info.tabId);
-		});
-
-		browser.tabs.onUpdated.addListener((tabId) => {
-			updateMenuItem(tabId);
-		});
-
-		pageTranslatorStateUpdatedHandler((state, tabId) => {
-			if (tabId === undefined || tabId !== this.$tabState.getState().tabId) return;
-
-			this.tabStateUpdated({
-				tabId,
-				isTranslating: state.isTranslated,
-			});
-		});
-	}
-
-	private onClickMenu = async (
-		info: browser.Menus.OnClickData,
-		tab: browser.Tabs.Tab | undefined,
-	) => {
-		if (info.menuItemId !== this.menuItemId) return;
-
-		const tabId = tab?.id;
-		if (tabId === undefined) return;
-
-		const isTranslating = this.$tabState.getState().isTranslating;
-
-		// Disable translating
-		if (isTranslating) {
-			disableTranslatePage(tabId);
-			return;
-		}
-
-		// Translate page
-		const config = await getConfig();
-		const translateState = await getPageTranslateState(tabId);
-
-		let from = translateState.translateDirection?.from ?? null;
-		if (from === null) {
-			const translatorFeatures = await getTranslatorFeatures();
-			from = translatorFeatures.isSupportAutodetect
-				? 'auto'
-				: translatorFeatures.supportedLanguages[0];
-		}
-
-		const to = translateState.translateDirection?.to ?? config.language;
-		enableTranslatePage(tabId, from, to);
-	};
-}
+import { TranslatePageContextMenu } from './ContextMenus/TranslatePageContextMenu';
 
 /**
  * Manage global states and application context
@@ -164,11 +55,6 @@ export class App {
 
 		await this.handleConfigUpdates();
 		await this.setupRequestHandlers();
-
-		// TODO: move to a `handleConfigUpdates` method
-		// TODO: enable only when option is enabled in config
-		const translatePageContextMenu = new TranslatePageContextMenu();
-		translatePageContextMenu.start();
 	}
 
 	private async setupRequestHandlers() {
@@ -226,5 +112,9 @@ export class App {
 				const isEnabled = enabled && mode === 'contextMenu';
 				toggleTranslateItemInContextMenu(isEnabled);
 			});
+
+		// TODO: enable only when option is enabled in config
+		const translatePageContextMenu = new TranslatePageContextMenu();
+		translatePageContextMenu.enable();
 	}
 }
